Silly fixes in CLI tooling, did not realize how many name changes I had made
