+++ tests/test_query.py
